Set CMake language variables when necessary
This adds explicit support for the "c", "cxx", and "python" package dependencies.  In the
event that "python" is used, the Python_ROOT CMake variable will be included as part of
the generated CMakePresets.json file, which influences the find_package(Python ...)
command.

Something to consider: we may need to also set the Python3_ROOT CMake variable so that
users can also type find_package(Python3 ...).